// src/app/models/indice.ts
export interface Indice {
  id: number;
  zona: number;
  quadra: string;
  qtdA1: number;
  qtdA2: number;
  qtdB: number;
  qtdC: number;
  qtdD1: number;
  qtdD2: number;
  qtdE: number;
  qtdOutroTpDeposito: number;
  qtdTspl: number;
  qtdTisa: number;
  qtdTvz: number;
  qtdTcpl: number;
  qtdTerrenoBaldio: number;
  qtdOutrTpImovel: number;
  breteau: number;
  ifp: number;
  risco: string;
  dataCriacao: Date;

}