// src/app/models/dado-paciente.ts
export class DadoPaciente {
  notificados: number;
  somaCasosPositivos: number;
  somaCasosEstrangeiros: number;
  somaCasosSuspeitos: number;
  internados: number;
  somaObitos: number;
  somaObitosSuspeitos: number;
  ubsId: number;
  ubsName: string;
  tipoArboviroseCount: { [key: string]: number };

}